Fix bug exercise3 + add global exception handling + documentation

+++ Wx.Exercises/Wx.Exercises.Services/Proxies/WxApiProxy/Models/Trolley.cs
@@ -26,7 +26,7 @@ namespace Wx.Exercises.Services.Proxies.WxApiProxy.Models
         [JsonProperty("quantities")]
         public List<TrolleyQuantity> Quantities { get; init; }
         [JsonProperty("total")]
-        public int Total { get; init; }
+        public double Total { get; init; }
     }
 
     public class TrolleyQuantity
